// peak-app/src/redux/slices/blog/types.ts
export interface BlogConfiguration {
    // Core
    id?: string
    title: string
    description: string
    subdomain: string

    // Configuration
    cover_image_url?: string
    fav_icon_url?: string
    twitter_url?: string
    instagram_url?: string
    linkedin_url?: string
    github_url?: string
}